+++ src/RanVec.hpp
@@ -13,8 +13,8 @@ class RanVec{
 	public:
 		RanVec(unsigned long int seed = 0);
 		
-		std::vector<double> get_rands(long unsigned int n);
-		std::vector<Point> get_vec_rands(long unsigned int n, int dim);
+		std::vector<double> get_rands(int n);
+		std::vector<Point> get_vec_rands(int n, int dim);
 		~RanVec(){gsl_rng_free(generator);}
 
 	private:
